In case of local mailserver installation change EMail field in Client Onboarding Request to read-only
This change is for the case when the local mailserver is set-up during solution deployment and active. In this case we can not send emails to external email addresses and changing the email field in the Client Onboarding Request is therefore misleading. To avoid this confusion we in this case set the EMail field in the Client Onboarding request  to read only. So after searching a Client when the fields in the Client Onboarding Request page 1 are automatically filled we now also set the email field to read-only.